+++ tortoise/fields/data.py
@@ -17,7 +17,13 @@ from pypika_tortoise.terms import Term
 
 from tortoise import timezone
 from tortoise.exceptions import ConfigurationError, FieldError
-from tortoise.fields.base import Field
+from tortoise.fields.base import (
+    Field,
+    FieldKwargs,
+    JSONFieldKwargs,
+    _FieldKwargsCommon,
+    _FieldKwargsNoPk,
+)
 from tortoise.timezone import get_default_timezone, get_timezone, get_use_tz, localtime
 from tortoise.validators import MaxLengthValidator
 
@@ -30,13 +36,22 @@ except ImportError:  # pragma: nocoverage
 
 try:
     from pydantic import BaseModel as _PydanticBaseModel
-    from pydantic._internal._model_construction import ModelMetaclass as _PydanticModelMetaclass
+    from pydantic._internal._model_construction import (
+        ModelMetaclass as _PydanticModelMetaclass,
+    )
 except ImportError:
     _PydanticBaseModel = None  # type: ignore[assignment,misc]
     _PydanticModelMetaclass = None  # type: ignore[assignment,misc]
 
 if TYPE_CHECKING:  # pragma: nocoverage
+    import sys
+
     from tortoise.models import Model
+
+    if sys.version_info >= (3, 11):
+        from typing import Unpack
+    else:  # pragma: no cover
+        from typing_extensions import Unpack
 
 __all__ = (
     "BigIntField",
@@ -107,7 +122,7 @@ class IntField(Field[T_INT], int):
         primary_key: bool | None = None,
         *,
         null: Literal[False] = False,
-        **kwargs: Any,
+        **kwargs: Unpack[_FieldKwargsNoPk],
     ) -> None: ...
 
     @overload
@@ -116,7 +131,7 @@ class IntField(Field[T_INT], int):
         primary_key: bool | None = None,
         *,
         null: Literal[True],
-        **kwargs: Any,
+        **kwargs: Unpack[_FieldKwargsNoPk],
     ) -> None: ...
 
     def __init__(self, primary_key: bool | None = None, **kwargs: Any) -> None:
@@ -222,12 +237,20 @@ class CharField(Field[T_STR]):
 
     @overload
     def __init__(
-        self: CharField[str], max_length: int, *, null: Literal[False] = False, **kwargs: Any
+        self: CharField[str],
+        max_length: int,
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
     def __init__(
-        self: CharField[str | None], max_length: int, *, null: Literal[True], **kwargs: Any
+        self: CharField[str | None],
+        max_length: int,
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, max_length: int, **kwargs: Any) -> None:
@@ -256,13 +279,35 @@ class CharField(Field[T_STR]):
             return f"NVARCHAR2({self.field.max_length})"
 
 
-class TextField(Field[str], str):  # type: ignore
+class TextField(Field[T_STR], str):  # type: ignore
     """
     Large Text field.
     """
 
     indexable = False
     SQL_TYPE = "TEXT"
+
+    @overload
+    def __init__(
+        self: TextField[str],
+        *,
+        primary_key: bool | None = None,
+        unique: bool = False,
+        db_index: bool = False,
+        null: Literal[False] = False,
+        **kwargs: Unpack[_FieldKwargsCommon],
+    ) -> None: ...
+
+    @overload
+    def __init__(
+        self: TextField[str | None],
+        *,
+        primary_key: bool | None = None,
+        unique: bool = False,
+        db_index: bool = False,
+        null: Literal[True],
+        **kwargs: Unpack[_FieldKwargsCommon],
+    ) -> None: ...
 
     def __init__(
         self,
@@ -315,12 +360,18 @@ class BooleanField(Field[T_BOOL]):
 
     @overload
     def __init__(
-        self: BooleanField[bool], *, null: Literal[False] = False, **kwargs: Any
+        self: BooleanField[bool],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
     def __init__(
-        self: BooleanField[bool | None], *, null: Literal[True], **kwargs: Any
+        self: BooleanField[bool | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None:
@@ -357,7 +408,7 @@ class DecimalField(Field[T_DECIMAL], Decimal):  # type: ignore
         decimal_places: int,
         *,
         null: Literal[False] = False,
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
@@ -367,7 +418,7 @@ class DecimalField(Field[T_DECIMAL], Decimal):  # type: ignore
         decimal_places: int,
         *,
         null: Literal[True],
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, max_digits: int, decimal_places: int, **kwargs: Any) -> None:
@@ -440,7 +491,7 @@ class DatetimeField(Field[T_DATETIME], datetime.datetime):
         auto_now_add: bool = False,
         *,
         null: Literal[False] = False,
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
@@ -450,7 +501,7 @@ class DatetimeField(Field[T_DATETIME], datetime.datetime):
         auto_now_add: bool = False,
         *,
         null: Literal[True],
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, auto_now: bool = False, auto_now_add: bool = False, **kwargs: Any) -> None:
@@ -530,12 +581,18 @@ class DateField(Field[T_DATE], datetime.date):
 
     @overload
     def __init__(
-        self: DateField[datetime.date], *, null: Literal[False] = False, **kwargs: Any
+        self: DateField[datetime.date],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
     def __init__(
-        self: DateField[datetime.date | None], *, null: Literal[True], **kwargs: Any
+        self: DateField[datetime.date | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None:
@@ -574,7 +631,7 @@ class TimeField(Field[T_TIME], datetime.time):
         auto_now_add: bool = False,
         *,
         null: Literal[False] = False,
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
@@ -584,7 +641,7 @@ class TimeField(Field[T_TIME], datetime.time):
         auto_now_add: bool = False,
         *,
         null: Literal[True],
-        **kwargs: Any,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, auto_now: bool = False, auto_now_add: bool = False, **kwargs: Any) -> None:
@@ -654,12 +711,18 @@ class TimeDeltaField(Field[T_TIMEDELTA]):
 
     @overload
     def __init__(
-        self: TimeDeltaField[datetime.timedelta], *, null: Literal[False] = False, **kwargs: Any
+        self: TimeDeltaField[datetime.timedelta],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
     def __init__(
-        self: TimeDeltaField[datetime.timedelta | None], *, null: Literal[True], **kwargs: Any
+        self: TimeDeltaField[datetime.timedelta | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None:
@@ -692,11 +755,19 @@ class FloatField(Field[T_FLOAT], float):
 
     @overload
     def __init__(
-        self: FloatField[float], *, null: Literal[False] = False, **kwargs: Any
+        self: FloatField[float],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
-    def __init__(self: FloatField[float | None], *, null: Literal[True], **kwargs: Any) -> None: ...
+    def __init__(
+        self: FloatField[float | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
+    ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(**kwargs)
@@ -748,7 +819,7 @@ class JSONField(Field[T], dict, list):  # type: ignore
         self,
         encoder: JsonDumpsFunc = JSON_DUMPS,
         decoder: JsonLoadsFunc = JSON_LOADS,
-        **kwargs: Any,
+        **kwargs: Unpack[JSONFieldKwargs],
     ) -> None:
         super().__init__(**kwargs)
         self.encoder = encoder
@@ -820,10 +891,20 @@ class UUIDField(Field[T_UUID], UUID):
         SQL_TYPE = "UUID"
 
     @overload
-    def __init__(self: UUIDField[UUID], *, null: Literal[False] = False, **kwargs: Any) -> None: ...
+    def __init__(
+        self: UUIDField[UUID],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
+    ) -> None: ...
 
     @overload
-    def __init__(self: UUIDField[UUID | None], *, null: Literal[True], **kwargs: Any) -> None: ...
+    def __init__(
+        self: UUIDField[UUID | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
+    ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None:
         if (kwargs.get("primary_key") or kwargs.get("pk", False)) and "default" not in kwargs:
@@ -852,12 +933,18 @@ class BinaryField(Field[T_BINARY], bytes):  # type: ignore
 
     @overload
     def __init__(
-        self: BinaryField[bytes], *, null: Literal[False] = False, **kwargs: Any
+        self: BinaryField[bytes],
+        *,
+        null: Literal[False] = False,
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     @overload
     def __init__(
-        self: BinaryField[bytes | None], *, null: Literal[True], **kwargs: Any
+        self: BinaryField[bytes | None],
+        *,
+        null: Literal[True],
+        **kwargs: Unpack[FieldKwargs],
     ) -> None: ...
 
     def __init__(self, **kwargs: Any) -> None: